requirements.txt
Django>=4.2
MetaTrader5
pandas
djangorestframework
Pillow
django-crispy-forms